ADD: 댓글 작성시 포인트 및 유저테이블 리뷰수 변동 함수

+++ src/main/java/hello/service/animations/ReviewService.java
@@ -7,6 +7,7 @@ import hello.entity.user.User;
 import hello.repository.db.AnimationRepository;
 import hello.repository.db.ReviewRepository;
 import hello.repository.db.UserRepository;
+import hello.service.basic.PointService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
@@ -23,8 +24,9 @@ public class ReviewService {
     private final ReviewRepository reviewRepository;
     private final AnimationRepository animationRepository;
     private final UserRepository userRepository;
+    private final PointService pointService;
 
-    public void addReview(AniReviewDTO aniReviewDTO) {
+    public void addReview(AniReviewDTO aniReviewDTO, User user) {
 
         Optional<Animation> animationOpt = animationRepository.findById(aniReviewDTO.getAnimationId());
         Optional<User> userOpt = userRepository.findById(aniReviewDTO.getUserId());
@@ -39,6 +41,16 @@ public class ReviewService {
 
             reviewRepository.save(review);
 
+            int currentReviewCount = user.getReviewCount();
+
+            if (currentReviewCount == 0) {
+                pointService.increasePoint(user, 3);
+            } else {
+                pointService.increasePoint(user, 1);
+            }
+
+            user.setReviewCount(currentReviewCount + 1);
+            userRepository.save(user);
         } else {
             System.out.println("Animation or User not found");
             if (animationOpt.isEmpty()) {
@@ -54,7 +66,6 @@ public class ReviewService {
     public Review findById(Long id) {
         return reviewRepository.findById(id).orElse(null);
     }
-
     public Review save(Review review) {
         return reviewRepository.save(review);
     }
@@ -62,7 +73,19 @@ public class ReviewService {
 
     // @ 리뷰 삭제
     public void deleteReview(long reviewId) {
-        reviewRepository.deleteById(reviewId);
+        Optional<Review> reviewOpt = reviewRepository.findById(reviewId);
+
+        if (reviewOpt.isPresent()) {
+            Review review = reviewOpt.get();
+            User user = review.getUser();
+
+            user.setReviewCount(user.getReviewCount() - 1);
+            userRepository.save(user);
+
+            reviewRepository.deleteById(reviewId);
+        } else {
+            System.out.println("Review not found for id: " + reviewId);
+        }
     }
 
     // 매일 자정 실행
